pyproject.toml
[project]
name = "event-agent"
version = "0.1.0"
description = "AI-Agent, der Familienveranstaltungen für einen Ort und Zeitraum findet."
requires-python = ">=3.11"
dependencies = [
    "strands-agents[openai]>=0.1.0",
    "tavily-python>=0.5.0",
    "ddgs>=9.0",
    "httpx>=0.27",
    "trafilatura>=1.12",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2",
    "python-dotenv>=1.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6",
]
